Monorepo POC
