+++ composer.json
@@ -8,14 +8,7 @@
         }
     ],
     "license": "GPL-2.0",
-    "config": {
-        "platform": {
-            "php": "7.3.0"
-        },
-        "process-timeout": 0
-    },
     "require": {
-        "php": ">=7.3.0",
         "ahand/mobileesp": "dev-master",
         "cap60552/php-sip2": "1.0.0",
         "colinmollenhour/credis": "1.12.1",
@@ -67,13 +60,14 @@
         "pear/http_request2": "2.5.0",
         "pear/validate_ispn": "0.8.0",
         "phing/phing": "2.17.0",
+        "php": ">=7.3.0",
         "ppito/laminas-whoops": "2.2.0",
         "scssphp/scssphp": "1.6.0",
         "serialssolutions/summon": "1.3.1",
         "slm/locale": "0.5.0",
+        "swagger-api/swagger-ui": "3.52.0",
         "symfony/console": "5.3.6",
         "symfony/yaml": "5.3.6",
-        "swagger-api/swagger-ui": "3.52.0",
         "vstelmakh/url-highlight": "3.0.0",
         "vufind-org/vufindcode": "1.2",
         "vufind-org/vufinddate": "1.0.0",
@@ -94,7 +88,24 @@
         "phpstan/phpstan": "1.2.0",
         "phpunit/phpunit": "9.5.10",
         "sebastian/phpcpd": "6.0.3",
-        "squizlabs/php_codesniffer": "3.6.2"
+        "squizlabs/php_codesniffer": "3.6.2",
+        "symplify/monorepo-builder": "^10.0"
+    },
+    "autoload": {
+        "psr-4": {
+            "VuFindView\\": "packages/vufindview/src/"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "VuFindView\\": "packages/vufindview/test/unit/"
+        }
+    },
+    "config": {
+        "platform": {
+            "php": "7.3.0"
+        },
+        "process-timeout": 0
     },
     "replace": {
         "laminas/laminas-cache-storage-adapter-apc": "*",
@@ -102,7 +113,8 @@
         "laminas/laminas-cache-storage-adapter-memcache": "*",
         "laminas/laminas-cache-storage-adapter-mongodb": "*",
         "laminas/laminas-cache-storage-adapter-wincache": "*",
-        "laminas/laminas-cache-storage-adapter-xcache": "*"
+        "laminas/laminas-cache-storage-adapter-xcache": "*",
+        "vufind/vufindview": "self.version"
     },
     "extra": {
         "merge-plugin": {
@@ -119,7 +131,10 @@
         }
     },
     "scripts": {
-        "phing-install-dependencies": ["phing patch-dependencies", "phing installsolr installswaggerui"],
+        "phing-install-dependencies": [
+            "phing patch-dependencies",
+            "phing installsolr installswaggerui"
+        ],
         "post-install-cmd": "@phing-install-dependencies",
         "post-update-cmd": "@phing-install-dependencies",
         "qa": "phing qa-console"
