+++ module/VuFind/tests/phpunit.xml
@@ -9,11 +9,13 @@
       <directory suffix=".php">../../VuFindDevTools/src</directory>
       <directory suffix=".php">../../VuFindSearch/src</directory>
       <directory suffix=".php">../../VuFindTheme/src</directory>
+      <directory suffix=".php">../../../packages/vufindview/src</directory>
     </include>
   </coverage>
   <testsuites>
     <testsuite name="AllUnitTests">
       <directory>../../*/tests/unit-tests/src</directory>
+      <directory>../../../packages/vufindview/test/unit</directory>
     </testsuite>
     <testsuite name="AllIntegrationTests">
       <directory>../../*/tests/integration-tests/src</directory>
