+++ module/VuFindTheme/src/VuFindTheme/View/Helper/ConcatTrait.php
@@ -46,191 +46,20 @@ use VuFindTheme\ThemeInfo;
 trait ConcatTrait
 {
     /**
-     * Returns true if file should not be included in the compressed concat file
+     * Get the filesystem path for the provided resource file path
      *
-     * @param stdClass $item Element object
+     * @param string $path Path
      *
-     * @return bool
+     * @return ?string
      */
-    abstract protected function isExcludedFromConcat($item);
-
-    /**
-     * Get the folder name and file extension
-     *
-     * @return string
-     */
-    abstract protected function getFileType();
-
-    /**
-     * Get the file path from the element object
-     *
-     * @param stdClass $item Element object
-     *
-     * @return string
-     */
-    abstract protected function getResourceFilePath($item);
-
-    /**
-     * Set the file path of the element object
-     *
-     * @param stdClass $item Element object
-     * @param string   $path New path string
-     *
-     * @return stdClass
-     */
-    abstract protected function setResourceFilePath($item, $path);
-
-    /**
-     * Get the minifier that can handle these file types
-     *
-     * @return minifying object like \MatthiasMullie\Minify\JS
-     */
-    abstract protected function getMinifier();
-
-    /**
-     * Add a content security policy nonce to the item
-     *
-     * @param stdClass $item Item
-     *
-     * @return void
-     *
-     * @SuppressWarnings(PHPMD.UnusedFormalParameter)
-     */
-    protected function addNonce($item)
+    protected function getResourceFileFilesystemPath(string $path): ?string
     {
-        // Default implementation does nothing
-    }
-
-    /**
-     * Set the file path of the link object
-     *
-     * @param stdClass $item Link element object
-     *
-     * @return string
-     *
-     * @SuppressWarnings(PHPMD.UnusedFormalParameter)
-     */
-    public function getType($item)
-    {
-        return 'default';
-    }
-
-    /**
-     * Should we use the asset pipeline to join files together and minify them?
-     *
-     * @var bool
-     */
-    protected $usePipeline = false;
-
-    /**
-     * Array of resource items by type, contains key as well
-     *
-     * @var array
-     */
-    protected $groups = [];
-
-    /**
-     * Future order of the concatenated file
-     *
-     * @var number
-     */
-    protected $concatIndex = null;
-
-    /**
-     * Check if config is enamled for this file type
-     *
-     * @param string|bool $config Config for current application environment
-     *
-     * @return bool
-     */
-    protected function enabledInConfig($config)
-    {
-        if ($config === false || $config == 'off') {
-            return false;
-        }
-        if ($config == '*' || $config == 'on'
-            || $config == 'true' || $config === true
-        ) {
-            return true;
-        }
-        $settings = array_map('trim', explode(',', $config));
-        return in_array($this->getFileType(), $settings);
-    }
-
-    /**
-     * Initialize class properties related to concatenation of resources.
-     * All of the elements to be concatenated into groups and
-     * and those that need to remain on their own special group 'other'.
-     *
-     * @return bool True if there are items
-     */
-    protected function filterItems()
-    {
-        $this->groups = [];
-        $groupTypes = [];
-
-        $this->getContainer()->ksort();
-
-        foreach ($this as $item) {
-            if ($this->isExcludedFromConcat($item)) {
-                $this->groups[] = [
-                    'other' => true,
-                    'item' => $item
-                ];
-                $groupTypes[] = 'other';
-                continue;
-            }
-
-            $path = $this->getFileType() . '/' . $this->getResourceFilePath($item);
-            $details = $this->themeInfo->findContainingTheme(
-                $path,
-                ThemeInfo::RETURN_ALL_DETAILS
-            );
-            // Deal with special case: $path was not found in any theme.
-            if (null === $details) {
-                $errorMsg = "Could not find file '$path' in theme files";
-                method_exists($this, 'logError')
-                    ? $this->logError($errorMsg) : error_log($errorMsg);
-                $this->groups[] = [
-                    'other' => true,
-                    'item' => $item
-                ];
-                $groupTypes[] = 'other';
-                continue;
-            }
-
-            $type = $this->getType($item);
-            $index = array_search($type, $groupTypes);
-            if ($index === false) {
-                $this->groups[] = [
-                    'items' => [$item],
-                    'key' => $details['path'] . filemtime($details['path'])
-                ];
-                $groupTypes[] = $type;
-            } else {
-                $this->groups[$index]['items'][] = $item;
-                $this->groups[$index]['key'] .=
-                    $details['path'] . filemtime($details['path']);
-            }
-        }
-
-        return count($groupTypes) > 0;
-    }
-
-    /**
-     * Get the path to the directory where we can cache files generated by
-     * this trait.
-     *
-     * @return string
-     */
-    protected function getResourceCacheDir()
-    {
-        if (!defined('LOCAL_CACHE_DIR')) {
-            throw new \Exception(
-                'Asset pipeline feature depends on the LOCAL_CACHE_DIR constant.'
-            );
-        }
-        return LOCAL_CACHE_DIR . '/public/';
+        $path = $this->getFileType() . '/' . $path;
+        $details = $this->themeInfo->findContainingTheme(
+            $path,
+            ThemeInfo::RETURN_ALL_DETAILS
+        );
+        return realpath($details['path']) ?? null;
     }
 
     /**
@@ -255,207 +84,7 @@ trait ConcatTrait
             return $urlHelper('home') . 'themes/' . $details['theme']
                 . '/' . $this->getFileType() . '/' . $path;
         }
-        // Locate/create concatenated asset file
-        $filename = md5($group['key']) . '.min.' . $this->getFileType();
-        // Minifier uses realpath, so do that here too to make sure we're not
-        // pointing to a symlink. Otherwise the path converter won't find the correct
-        // shared directory part.
-        $concatPath = realpath($this->getResourceCacheDir()) . '/' . $filename;
-        if (!file_exists($concatPath)) {
-            $lockfile = "$concatPath.lock";
-            $handle = fopen($lockfile, 'c+');
-            if (!is_resource($handle)) {
-                throw new \Exception("Could not open lock file $lockfile");
-            }
-            if (!flock($handle, LOCK_EX)) {
-                fclose($handle);
-                throw new \Exception("Could not lock file $lockfile");
-            }
-            // Check again if file exists after acquiring the lock
-            if (!file_exists($concatPath)) {
-                try {
-                    $this->createConcatenatedFile($concatPath, $group);
-                } catch (\Exception $e) {
-                    flock($handle, LOCK_UN);
-                    fclose($handle);
-                    throw $e;
-                }
-            }
-            flock($handle, LOCK_UN);
-            fclose($handle);
-        }
 
-        return $urlHelper('home') . 'cache/' . $filename;
-    }
-
-    /**
-     * Create a concatenated file from the given group of files
-     *
-     * @param string $concatPath Resulting file path
-     * @param array  $group      Object containing 'key' and stdobj file 'items'
-     *
-     * @throws \Exception
-     * @return void
-     */
-    protected function createConcatenatedFile($concatPath, $group)
-    {
-        $data = [];
-        foreach ($group['items'] as $item) {
-            $details = $this->themeInfo->findContainingTheme(
-                $this->getFileType() . '/'
-                . $this->getResourceFilePath($item),
-                ThemeInfo::RETURN_ALL_DETAILS
-            );
-            $details['path'] = realpath($details['path']);
-            $data[] = $this->getMinifiedData($details, $concatPath);
-        }
-        // Separate each file's data with a new line so that e.g. a file
-        // ending in a comment doesn't cause the next one to also get commented out.
-        file_put_contents($concatPath, implode("\n", $data));
-    }
-
-    /**
-     * Get minified data for a file
-     *
-     * @param array  $details    File details
-     * @param string $concatPath Target path for the resulting file (used in minifier
-     * for path mapping)
-     *
-     * @throws \Exception
-     * @return string
-     */
-    protected function getMinifiedData($details, $concatPath)
-    {
-        if ($this->isMinifiable($details['path'])) {
-            $minifier = $this->getMinifier();
-            $minifier->add($details['path']);
-            $data = $minifier->execute($concatPath);
-        } else {
-            $data = file_get_contents($details['path']);
-            if (false === $data) {
-                throw new \Exception(
-                    "Could not read file {$details['path']}"
-                );
-            }
-        }
-        return $data;
-    }
-
-    /**
-     * Process and return items in index order
-     *
-     * @param string|int $indent Amount of whitespace/string to use for indention
-     *
-     * @return string
-     */
-    protected function outputInOrder($indent)
-    {
-        // Some of this logic was copied from HeadScript; it does not all apply
-        // when incorporated into HeadLink, but it has no harmful side effects.
-        $indent = (null !== $indent)
-            ? $this->getWhitespace($indent)
-            : $this->getIndent();
-
-        if ($this->view) {
-            $useCdata = $this->view->plugin('doctype')->isXhtml();
-        } else {
-            $useCdata = $this->useCdata;
-        }
-
-        $escapeStart = ($useCdata) ? '//<![CDATA[' : '//<!--';
-        $escapeEnd   = ($useCdata) ? '//]]>' : '//-->';
-
-        $output = [];
-        foreach ($this->groups as $group) {
-            if (isset($group['other'])) {
-                $output[] = $this->itemToString(
-                    $group['item'],
-                    $indent,
-                    $escapeStart,
-                    $escapeEnd
-                );
-            } else {
-                // Note that we  use parent::itemToString() below instead of
-                // $this->itemToString() to bypass VuFind logic that determines
-                // file paths within the theme (not appropriate for concatenated
-                // files, which are stored in a theme-independent cache).
-                $path = $this->getConcatenatedFilePath($group);
-                $item = $this->setResourceFilePath($group['items'][0], $path);
-                $this->addNonce($item);
-                $output[] = parent::itemToString(
-                    $item,
-                    $indent,
-                    $escapeStart,
-                    $escapeEnd
-                );
-            }
-        }
-
-        return $indent . implode(
-            $this->escape($this->getSeparator()) . $indent,
-            $output
-        );
-    }
-
-    /**
-     * Check if a file is minifiable i.e. does not have a pattern that denotes it's
-     * already minified
-     *
-     * @param string $filename File name
-     *
-     * @return bool
-     */
-    protected function isMinifiable($filename)
-    {
-        $basename = basename($filename);
-        return preg_match('/\.min\.(js|css)/', $basename) === 0;
-    }
-
-    /**
-     * Can we use the asset pipeline?
-     *
-     * @return bool
-     */
-    protected function isPipelineActive()
-    {
-        if ($this->usePipeline) {
-            try {
-                $cacheDir = $this->getResourceCacheDir();
-            } catch (\Exception $e) {
-                $this->usePipeline = $cacheDir = false;
-                error_log($e->getMessage());
-            }
-            if ($cacheDir && !is_writable($cacheDir)) {
-                $this->usePipeline = false;
-                error_log("Cannot write to $cacheDir; disabling asset pipeline.");
-            }
-        }
-        return $this->usePipeline;
-    }
-
-    /**
-     * Render link elements as string
-     * Customized to minify and concatenate
-     *
-     * @param string|int $indent Amount of whitespace or string to use for indention
-     *
-     * @return string
-     */
-    public function toString($indent = null)
-    {
-        // toString must not throw exception
-        try {
-            if (!$this->isPipelineActive() || !$this->filterItems()
-                || count($this) == 1
-            ) {
-                return parent::toString($indent);
-            }
-
-            return $this->outputInOrder($indent);
-        } catch (\Exception $e) {
-            error_log($e->getMessage());
-        }
-
-        return '';
+        return parent::getConcatenatedFilePath($group);
     }
 }
