+++ monorepo-builder.php
@@ -1,0 +1,31 @@
+<?php
+// @codingStandardsIgnoreStart
+declare(strict_types=1);
+
+use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
+use Symplify\MonorepoBuilder\Release\ReleaseWorker\PushNextDevReleaseWorker;
+use Symplify\MonorepoBuilder\Release\ReleaseWorker\PushTagReleaseWorker;
+use Symplify\MonorepoBuilder\Release\ReleaseWorker\SetCurrentMutualDependenciesReleaseWorker;
+use Symplify\MonorepoBuilder\Release\ReleaseWorker\SetNextMutualDependenciesReleaseWorker;
+use Symplify\MonorepoBuilder\Release\ReleaseWorker\TagVersionReleaseWorker;
+use Symplify\MonorepoBuilder\Release\ReleaseWorker\UpdateBranchAliasReleaseWorker;
+use Symplify\MonorepoBuilder\ValueObject\Option;
+// @codingStandardsIgnoreEnd
+
+return static function (ContainerConfigurator $containerConfigurator): void {
+    $parameters = $containerConfigurator->parameters();
+    $parameters->set(Option::PACKAGE_DIRECTORIES, [__DIR__ . '/packages']);
+    $parameters->set(
+        Option::PACKAGE_DIRECTORIES_EXCLUDES,
+        [__DIR__ . '/packages/DEBIAN']
+    );
+    $parameters->set(Option::DEFAULT_BRANCH_NAME, 'monorepo-release');
+
+    $services = $containerConfigurator->services();
+    $services->set(SetCurrentMutualDependenciesReleaseWorker::class);
+    $services->set(TagVersionReleaseWorker::class);
+    $services->set(PushTagReleaseWorker::class);
+    $services->set(SetNextMutualDependenciesReleaseWorker::class);
+    $services->set(UpdateBranchAliasReleaseWorker::class);
+    $services->set(PushNextDevReleaseWorker::class);
+};
