+++ packages/vufindview/composer.json
@@ -1,0 +1,21 @@
+{
+    "name": "vufind/vufindview",
+    "autoload": {
+        "psr-4": {
+            "VuFindView\\": "src/"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "VuFindView\\": "test/unit/"
+        }
+    },
+    "require": {
+        "matthiasmullie/minify": "1.3.66",
+        "laminas/laminas-servicemanager": "3.7.0",
+        "laminas/laminas-view": "2.15.0"
+    },
+    "require-dev": {
+        "phpunit/phpunit": "9.5.10"
+    }
+}
