+++ tests/phpcs.xml
@@ -2,8 +2,11 @@
 <ruleset name="VuFind Coding Standards" namespace="VuFind\PHPCS">
   <description>Coding standards for VuFind.</description>
   <file>../module</file>
+  <file>../packages</file>
   <exclude-pattern>module/[^/]+/config/*</exclude-pattern>
   <exclude-pattern>module/[^/]+/tests/*</exclude-pattern>
+  <exclude-pattern>packages/DEBIAN</exclude-pattern>
+  <exclude-pattern>packages/[^/]+/test/*</exclude-pattern>
   <arg name="extensions" value="php"/>
   <rule ref="PEAR">
     <exclude name="PEAR.Commenting.FunctionComment.ParamCommentAlignment" />
